requirements.txt
mitmproxy
playwright
requests
PyJWT
PySide6
google-generativeai
fpdf2
stem
requests[socks]
